Consolidate FLAC fixture helpers into fuzz_check::fixtures (#139)
Three copies (format tests/common, core tests/common, fuzz_check's
private set) become one. The self-dev-dependency turns the fuzzing
feature on for musefs-format's own test builds, so the proptests now
run under a plain cargo test -p musefs-format.
