+++ musefs-core/tests/common/mod.rs
@@ -5,47 +5,9 @@ use std::path::Path;
 pub mod corpus;
 pub mod report;
 
-pub fn flac_block(block_type: u8, body: &[u8], is_last: bool) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.push((if is_last { 0x80 } else { 0 }) | (block_type & 0x7F));
-    let len = body.len();
-    out.push(((len >> 16) & 0xFF) as u8);
-    out.push(((len >> 8) & 0xFF) as u8);
-    out.push((len & 0xFF) as u8);
-    out.extend_from_slice(body);
-    out
-}
-
-pub fn streaminfo_body() -> Vec<u8> {
-    let mut b = vec![
-        0x10, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0A, 0xC4, 0x42, 0xF0, 0x00,
-        0x00, 0x00, 0x00,
-    ];
-    b.extend_from_slice(&[0u8; 16]);
-    b
-}
-
-pub fn vorbis_comment_body(vendor: &str, comments: &[&str]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(&(vendor.len() as u32).to_le_bytes());
-    out.extend_from_slice(vendor.as_bytes());
-    out.extend_from_slice(&(comments.len() as u32).to_le_bytes());
-    for c in comments {
-        out.extend_from_slice(&(c.len() as u32).to_le_bytes());
-        out.extend_from_slice(c.as_bytes());
-    }
-    out
-}
-
-pub fn make_flac(blocks: &[(u8, Vec<u8>)], audio: &[u8]) -> Vec<u8> {
-    let mut out = Vec::new();
-    out.extend_from_slice(b"fLaC");
-    for (i, (bt, body)) in blocks.iter().enumerate() {
-        out.extend_from_slice(&flac_block(*bt, body, i == blocks.len() - 1));
-    }
-    out.extend_from_slice(audio);
-    out
-}
+pub use musefs_format::fuzz_check::fixtures::{
+    flac_block, make_flac, streaminfo_body, vorbis_comment_body,
+};
 
 /// Write a simple FLAC (STREAMINFO + comment + audio) to `path`,
 /// returning (audio_offset, audio_length).
