+++ musefs-format/Cargo.toml
@@ -23,6 +23,10 @@ crc = "3"
 hound = "3"
 musefs-db = { path = "../musefs-db", version = "0.2.0" }
 proptest = "1"
+# Self-dependency: turns the `fuzzing` feature on for this crate's own test
+# builds, so tests/ can use fuzz_check and the proptests run without an
+# explicit --features flag.
+musefs-format = { path = ".", features = ["fuzzing"] }
 
 [lints]
 workspace = true
